pyproject.toml
[project]
name = "paymentssanic"
version = "0.1.0"
description = "Async REST API for managing users, accounts and payments"
authors = [
    {name = "YevgeniyKapustin",email = "kapustaevg@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1,<2.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "email-validator>=2.2.0,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "psycopg[binary]>=3.2.6,<4.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "sanic>=24.12.0,<25.0.0",
    "sanic-ext>=24.12.0,<25.0.0",
    "sqlalchemy>=2.0.39,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.21.0,<0.22.0",
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.26.0,<1.0.0",
    "pytest-cov>=5.0.0,<7.0.0",
    "ruff>=0.11.0,<1.0.0",
    "black>=25.1.0,<26.0.0",
    "mypy>=1.15.0,<2.0.0",
    "isort>=5.13.0,<6.0.0",
    "bandit>=1.8.0,<2.0.0",
    "safety>=3.2.0,<4.0.0",
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".poetry_cache", "alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=xml --cov-fail-under=80"
filterwarnings = [
    "ignore:websockets\\.connection was renamed to websockets\\.protocol.*:DeprecationWarning",
    "ignore:The HMAC key is .* below the minimum recommended length.*:jwt.warnings.InsecureKeyLengthWarning",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
py_version = 312
skip_gitignore = true

[tool.mypy]
python_version = "3.12"
plugins = []
ignore_missing_imports = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
pretty = true
exclude = ["alembic/versions/", ".*types.py$", "tests/"]

[[tool.mypy.overrides]]
module = [
    "sqlalchemy.*",
    "httpx.*",
    "passlib.*",
    "pydantic_settings",
    "sanic_ext",
    "alembic",
    "types",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sanic", "sanic.*", "jwt", "jwt.*"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "alembic"]
skips = ["B101"]
